Update patch set 6
Patch Set 6:

(4 comments)

It is a good idea to set up Eclipse with formatting, Sonar and CheckStyle, see https://wiki.onap.org/display/DW/Setting+Up+Your+Development+Environment#SettingUpYourDevelopmentEnvironment-IDE:Eclipse:SettingupEclipse.

Now there are some lines in the code which are longer that 120 characters. A recommendation is to not have lines longer than 120 characters, for better readability. If formatting, Sonar and CheckStyle are set up, these and other issues will be handled automatically or shown in Eclipse so that they can be avoided before committing.

Patch-set: 6
Label: Code-Review=0