Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)

We removed this part of the code due to duplicate logs, the same information is printed moments earlier in the RequestSender class in the getResult method.
Yes, the main goal was to hide the password but we want also improve code and get rid of unnecessary duplicated logs.

Patch-set: 2