Update patch set 11
Patch Set 11:

> Patch Set 11: Code-Review-1
> 
> (1 comment)
Hi, Joseph,
we just to check the original ves message to check whether it is need to do kpi computation, and no matter whether the ves message to do kpi computation or not, the original ves message  will still publish to MR，and it does not influence the already existing function of the pm-mapper. if the ves message kpi related, it is only to do simple computation and generate a kpi ves message to publish MR which is the same topic of PERFORMANCE_MEASUREMENTS。

and for this release, I hope we can do this like it. and later we can consider it as a independent feature.

Patch-set: 11