Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> I'd like a demo to see how you can deploy the service with the previous existing functionality ONLY in operation, and then see how you can enable the new functionality.

there is no diff between the new pm-mapper and the previous pm-mapper for deploying, installing and operations, for new pm-mapper, the change is to add a kpi_config property in pm-mapper blueprint, and kpi computation will do computation based on the configuration. Before pm-mapper publish ves data to MR, pm-mapper add a logic to check whether the ves data should do kpi computation(based on the eventName and measTypes defined in the cbs configuration), if so, will do kpi computation and generate the ves result and publish to MR(the same topic). and whether the ves data will do kpi computation or not, the original will still publish to MR

Patch-set: 11