Update patch set 12
Patch Set 12:

> Patch Set 11:
> 
> > Patch Set 11: Code-Review-1
> > 
> > (1 comment)
> Hi, Joseph,
> we just to check the original ves message to check whether it is need to do kpi computation, and no matter whether the ves message to do kpi computation or not, the original ves message  will still publish to MR，and it does not influence the already existing function of the pm-mapper. if the ves message kpi related, it is only to do simple computation and generate a kpi ves message to publish MR which is the same topic of PERFORMANCE_MEASUREMENTS。
> 
> and for this release, I hope we can do this like it. and later we can consider it as a independent feature.

I think this a design fault, I don't think we should merge that much technical debt.
I wasn't saying to move it into its own application or anything, just to move the KPI compute code into its own class, that way it leaves the original publishing mechanism in tact.
I don't think its a good idea to replace a method that was originally intended just to publish to MR, with a method designed to publish to MR but first do a lot of computation on it. it really defeats the original intent for this method.

Patch-set: 12
Reviewer: Gerrit User 4023 <4023@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>