Update patch set 6
Patch Set 6:

> Patch Set 5:
> 
> recheck

This time I found that a new error came "ERROR: (copyright-year-not-found-in-file) Copyright 2023 not found" in the pipeline, I changed copyright year to 2023 in 3 files and now that error is not happening, All testcases both unit and robot have passed successfully. Build is failing due to python version issue again. seems like it might take more time for robot environment issues to be fixed,Can you check the code and provide your comments?

Patch-set: 6